If javascript is unavailable display an HTML table with required data
This demonstrates the approach to be used to ensure the application is
accessable to screen reader users.

The view will parse the data passed to it in two ways it will generate
an HTML table which will be displayed by default and store the data as a
javascript variable.

If javascript is available the variable is parsed by the associated javascript
file the chart is generated and the table is removed.