Add aria compatable tags to the compare utilisation HTML
This will communicate to aria compatable readers that the div and all
sub-sections shouldn't be rendered by the screen reader. This is the case
because all of the content is presented in a much more accessible format i.e.
the HTML table.

These were identified using the following pa11y command:
```
pa11y http://localhost:3000/compare-utilisation --hide-elements 'div[role=presentation]'
```