If javascript is unavailable display an HTML table with required data (#14)
* If javascript is unavailable display an HTML table with required data

This demonstrates the approach to be used to ensure the application is
accessable to screen reader users.

The view will parse the data passed to it in two ways it will generate
an HTML table which will be displayed by default and store the data as a
javascript variable.

If javascript is available the variable is parsed by the associated javascript
file the chart is generated and the table is removed.

* Get utilisation comparison data from database

Replace the mocked out dataset originally used to prove the accessibility
capability with database data.

In order to do this the database structure and seed data needed updated.

* Hide non-javascript elements onload

Use the `js-enabled` class in combination with a new `non-javascript` class to
hide elements while the DOM is rendering. This prevents non-javascript only
elements flashing onto the screen while the javascript excutes and before they
are removed and replaced.

* Add aria compatable tags to the compare utilisation HTML

This will communicate to aria compatable readers that the div and all
sub-sections shouldn't be rendered by the screen reader. This is the case
because all of the content is presented in a much more accessible format i.e.
the HTML table.

These were identified using the following pa11y command:
```
pa11y http://localhost:3000/compare-utilisation --hide-elements 'div[role=presentation]'
```